+++ internal/terminal/terminal_flush_darwin.go
@@ -1,0 +1,8 @@
+package terminal
+
+import "golang.org/x/sys/unix"
+
+func flushTerminalInput(fd int) error {
+	const readQueue = 1
+	return unix.IoctlSetPointerInt(fd, unix.TIOCFLUSH, readQueue)
+}