+++ internal/terminal/terminal_flush_linux.go
@@ -1,0 +1,7 @@
+package terminal
+
+import "golang.org/x/sys/unix"
+
+func flushTerminalInput(fd int) error {
+	return unix.IoctlSetInt(fd, unix.TCFLSH, unix.TCIFLUSH)
+}